Create better_example.cc
The example in the video has a class with members which are "trivially movable", i.e. copied when moved. This is not the most instructive example!
This version has a class with an std::string member, and moving it has a different result from copying it. The output also gives better information.
I have kept the original example available and not overwritten it, to avoid confusion.
If you would like a little challenge, change the argument of the move constructor to const Test&&, and see if you can explain the result!